moved recenter for generic use
