+++ safe-route/src/components/mapcontainer.js
@@ -1,19 +1,10 @@
-import {React, useState, useEffect } from 'react';
+import { React, useState, useEffect } from 'react';
 import 'leaflet/dist/leaflet.css';
 import { MapContainer, TileLayer, Marker, useMap } from 'react-leaflet';
 import { currentLocIcon } from './icons/loc-icon';
+import Recenter from './recenter';
 
-
-const Recenter = ({lat, lng}) => {
-  const map = useMap();
-
-  useEffect(() => {
-      map.setView([lat, lng]);
-  }, [lat, lng]);
-  return null;
-}
-
-const MapCont = ({lat, lng}) => {
+const MapCont = ({ lat, lng }) => {
 
   return (
     <MapContainer center={[lat, lng]} zoom={12} style={{ width: "100%", height: "100vh" }}>
